seperate project editor into two fragments
