--- src/main/kotlin/org/wycliffeassociates/otter/jvm/app/ui/projecteditor/ChapterContext.kt
@@ -1,7 +1,0 @@
-package org.wycliffeassociates.otter.jvm.app.ui.projecteditor
-
-enum class ChapterContext {
-    RECORD,
-    VIEW_TAKES,
-    EDIT_TAKES
-}
