+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/app/ui/viewcontent/viewmodel/ViewContentViewModel.kt
@@ -1,0 +1,95 @@
+package org.wycliffeassociates.otter.jvm.app.ui.viewcontent.viewmodel
+
+import com.github.thomasnield.rxkotlinfx.changes
+import com.github.thomasnield.rxkotlinfx.observeOnFx
+import com.github.thomasnield.rxkotlinfx.toObservable
+import io.reactivex.Observable
+import javafx.beans.property.SimpleBooleanProperty
+import javafx.beans.property.SimpleObjectProperty
+import javafx.collections.FXCollections
+import javafx.collections.ObservableList
+import org.wycliffeassociates.otter.common.data.model.Collection
+import org.wycliffeassociates.otter.common.data.model.Content
+import org.wycliffeassociates.otter.common.domain.content.AccessTakes
+import org.wycliffeassociates.otter.jvm.app.ui.inject.Injector
+import tornadofx.*
+
+class ViewContentViewModel: ViewModel() {
+
+    private val injector: Injector by inject()
+    private val contentRepository = injector.contentRepository
+    private val takeRepository = injector.takeRepository
+
+    // The selected project
+    private var activeProject: Collection by property()
+    val activeProjectProperty = getProperty(ViewContentViewModel::activeProject)
+
+    // Selected child
+    private var activeCollection: Collection by property()
+    val activeCollectionProperty = getProperty(ViewContentViewModel::activeCollection)
+
+    private var activeContent: Content by property()
+    val activeContentProperty = getProperty(ViewContentViewModel::activeContent)
+
+    // List of content to display on the screen
+    // Boolean tracks whether the content has takes associated with it
+    val allContent: ObservableList<Pair<SimpleObjectProperty<Content>, SimpleBooleanProperty>>
+            = FXCollections.observableArrayList()
+    val filteredContent: ObservableList<Pair<SimpleObjectProperty<Content>, SimpleBooleanProperty>>
+            = FXCollections.observableArrayList()
+
+    // Whether the UI should show the plugin as active
+    private var showPluginActive: Boolean by property(false)
+    val showPluginActiveProperty = getProperty(ViewContentViewModel::showPluginActive)
+
+    private var loading: Boolean by property(false)
+    val loadingProperty = getProperty(ViewContentViewModel::loading)
+
+    val chapterModeEnabledProperty = SimpleBooleanProperty(false)
+    private val accessTakes = AccessTakes(contentRepository, takeRepository)
+
+    init {
+        activeCollectionProperty.toObservable().subscribe { selectChildCollection(it) }
+        Observable.merge(chapterModeEnabledProperty.toObservable(), allContent.changes()).subscribe { _ ->
+            filteredContent.setAll(
+                    if (chapterModeEnabledProperty.value == true) {
+                        allContent.filtered { it.first.value?.labelKey == "chapter" }
+                    } else {
+                        allContent.filtered { it.first.value?.labelKey != "chapter" }
+                    }
+            )
+        }
+    }
+
+    fun selectChildCollection(child: Collection) {
+        activeCollection = child
+        // Remove existing content so the user knows they are outdated
+        allContent.clear()
+        loading = true
+        contentRepository
+                .getByCollection(child)
+                .flatMapObservable {
+                    Observable.fromIterable(it)
+                }
+                .flatMapSingle { content ->
+                    accessTakes
+                            .getTakeCount(content)
+                            .map { Pair(content.toProperty(), SimpleBooleanProperty(it > 0)) }
+                }
+                .toList()
+                .observeOnFx()
+                .subscribe { retrieved ->
+                    retrieved.sortBy { it.first.value.sort }
+                    allContent.clear() // Make sure any content that might have been added are removed
+                    allContent.addAll(retrieved)
+                    loading = false
+                }
+    }
+
+    fun viewContentTakes(content: Content) {
+        // Launch the select takes page
+        // Might be better to use a custom scope to pass the data to the view takes page
+        activeContent = content
+//        workspace.dock<ViewTakesView>()
+    }
+}
